swupdate: call cml1_do_configure() at least once while configuring
Without this call, swupdate will be built with a default configuration which
is incompatible with our package format. Even though this function is already
called in the original recipe, it was done in the package's "git" directory.
Make sure the function is called at least once so that our defconfig is used in
the build.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>