install_linux_fw_uuu: reconfigure fastboot after partitioning
Fastboot is configured on boot with the current MMC configuration.
If we create a GPT partition table, we need to restart fastboot before we
try to program any partition with the 'flash' fastboot command.

Set 'fastboot_dev' to a different device, 'sata' for example an back to 'mmc'
to trigger a fastboot restart after creating the partition table on blank
devices.

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>