u-boot-dey: build cc6 U-Boot from the HEAD of v2015.04/maint
We were building the cc6 U-Boot from the dub-2015.04-r10.2 tag because later
versions assume the kernel image type to be zImage by default, while the cc6
still uses uImage in DEY 2.2.

However, this can be remedied by changing an environment variable. Do so in the
bootscript and build from the HEAD of v2015.04/maint to incorporate the latest
changes, including CC6N support and general fixes from the past few years. This
avoids having to maintain different U-Boot tags in our legacy DEY releases as
well as having to incorporate changes via Yocto recipe patches.

While at it, change the bootscript so that it uses a valid default fdt_file if
there's no board ID/variant programmed on the target.

https://onedigi.atlassian.net/browse/DEL-7661
https://onedigi.atlassian.net/browse/DEL-7672

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>