Revert "trustfence: change RAM location to the same values used in u-boot"
The AHAB decryption process takes the encrypted file from the address defined
in U-Boot and decrypts it into the address defined in this script. If both
addresses are the same, the decryption process ends up failing. This
happens even for signed-only images.

Maintain the original addresses in this script so they do not collide.

This reverts commit c970d87d5a0370b1c76e7e36e0563e2120ec1b9d.

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>