hardknott: imx-boot: sync with meta-imx hardknott-5.10.72-2.2.0 branch
NXP now uses the meta-freescale version of the imx-boot recipe as a base, and
applies a .bbappend in meta-imx to apply more customizations. Reflect this
in meta-digi-arm and merge said .bbappend with ours. While at it, omit one of
NXP's operations in their do_deploy, since it breaks our build and doesn't look
essential.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>