weston-init: change target from "graphical" to "multi-user"
In older Yocto versions, weston was launched via a udev rule, but now it gets
launched as part of the "graphical" system target. Since this target isn't
reached automatically on boot, the weston desktop will only appear if you
manually launch said target (or the weston service itself).

To make the desktop launch automatically, change the target to "multi-user",
which is the default target laucnhed on boot.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>