linux-firmware: remove SDMA packages
This makes it so bitbake obtains the SDMA firmware from the firmware-imx recipe
instead.

This also avoids including all of the linux-firmware packages in i.MX6 SDK
builds, preventing errors when building the ccimx6sbc SDK (linux-firmware
provides the community versions of the ath6k and ar3k firmwares, which clash
with our versions).

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>