weston-init: restore profile script to previous implementation
This is a partial revert of commit 3819ee66726ea9606755495662efc2184ec764c2.
NXP simplified the profile script when updating to hardknott-5.10.72-2.2.0, but
since the /run/user/0 directory isn't automatically created in sysvinit
environments, the weston desktop fails to start in these cases. The previous
version of the script automatically creates said directory if it's missing, so
use that version instead.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>