cog: adapt patches to v0.12.0
Now that the "--platform" parameter is optional, we don't need to hardcode the
platform anymore, so remove the first patch.

In this version, the FDO platform has been renamed to WL, and since the default
PACKAGECONFIG is meant to work with older cog versions, we need to manually
append the "wl" PACKAGECONFIG to build the wayland platform.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>