linux-dey: move the version specific info to the proper bitbake recipes
The license checksum and the source branch should be on the properly
versioned bitbake recipes. Otherwise, when we include this common file
in a different version linux recipe (for example, the upcoming 5.10) it
will inherit incorrect values.

While on it, fix the incorrect license checksum values for the linux 5.4
recipes.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>