trustfence: add Trustfence tools in default toolchain
In order to perform the standalone signature process, it was required
to rebuild the Toolchain with Trustfence support enabled.

CST source code is now available for downloading in the Digi FTP, so add
Trustfence sign scripts and cst/srktool to the default toolchain for it
to be used for standalone signature without rebuilding.

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>