kirkstone migration: update wpa-supplicant package
Mainly, refresh patches so they apply on top of wpa-supplicant 2.10

Murata patches from release: cypress-fmac-v5.10.9-2022_0331

NOTE: the wpa-supplicant patches in the cypress release apply on top of
1759a8e3f36a commit of the hostapd.git repository. That repository
includes both the wpa-supplicant and the hostapd packages code. In Yocto
the wpa-supplicant recipe uses the release tarball, so there is no
'hostapd' folder included. Thus, the cypress patches that changed code
in 'hostapd' and 'wpa_supplicant' folders, have been splitted in two
(the hostapd changes and the rest). In the recipe the patches that
changed 'hostapd' code have been marked as 'apply=no' because they won't
apply to the tarball, but they have not been deleted for traceability.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>