kirkstone migration: fix kernel-module-qualcomm build
Yocto is becoming more strict on not using tags in the SRCREV, as that
may lead to problems in the different tasks of a bitbake recipe.

As of Yocto 4.0, bitbake errors when such condition happens, with:

"Recipe uses a floating tag/branch without a fixed SRCREV"

So convert the SRCREV of all tags to the proper SHA1 revision.

Also, extend the 'fix-build-issues' to delete '-Werror' from the
compilation flags. Newer versions of GCC (as the one used in Yocto 4.0)
throw more (new) warnings, and the '-Werror' flag was making the build
to fail.

cc1: all warnings being treated as errors
lim_api.c:1057:17: error: this 'if' clause does not guard... [-Werror=misleading-indentation]
lim_admit_control.c:169:17: error: this 'if' clause does not guard... [-Werror=misleading-indentation]

Signed-off-by: Javier Viguera <javier.viguera@digi.com>