meta-digi-dey: remove obsolete python2 based recipe
This is the only python2-based recipe in our layer, so remove it to be
able to remove the 'meta-python2' layer from DEY as well.

Apparently it was introduced for some Optee dependence and it was used
in previous versions of 'meta-imx', but 'meta-imx' removed it as well
for Kirkstone, so we do the same here.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>