linux-dey: fix kernel build configuration
Commit 8cba3aaefb in 'meta-freescale' changed the kernel_localversion
function we are using in our kernel recipe. This leads to a not-properly
configured kernel build, where CONFIG_MODULES is disabled. That
config option is needed to build external kernel modules in other
recipes, so those recipes (kernel-module-qualcomm, cryptodev-module,
etc) fail to build.

The commit makes sure the kernel is properly configured for building by
extending the do_configure function in a "prepend".

Signed-off-by: Javier Viguera <javier.viguera@digi.com>