linux-dey: get rid of LOG kernel mirror
Use the canonical Stash repository to avoid synchronization problems
between the repo and the mirror.

Moreover the LOG mirror is hosted in an old VM machine with limited
specs, and although the transfer of the objects is faster, the counting
of the objects to transfer (which is done in the server) takes ages to
complete, so at the end, there is no time gain using the mirror.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>