meta-digi-arm: add FIPS supported wolfssl recipe
We don't want the FIPS version of wolfssl to be used by default, so
the recipe is configured as:

DEFAULT_PREFERENCE = "-1"

Unfortunately, DEFAULT_PREFERENCE has lower precedence than the layer
priority, so we need to put the recipe in a layer with priority less
or equal to the other layers having wolfssl recipes (meta-openembedded,
meta-wolfssl). This means in our case using 'meta-digi-arm' instead of
'meta-digi-dey', in which conceptually would fit better.

https://onedigi.atlassian.net/browse/DEL-8036

Signed-off-by: Javier Viguera <javier.viguera@digi.com>