meta-digi-dey: dey-examples: fix start/stop script of "connectcore-demo-example" service
The start/stop script of the service was using "killproc" and "pidofproc" functions from
"/etc/init.d/functions" to find the process PID. These functions rely on "pidof", which does
not work very well with Python scripts. Instead, use "pkill" and "pgrep" which allow to
search the executable in the full command line to retrieve the correct PID with the -f modifier.

Signed-off-by: David Escalona <david.escalona@digi.com>