usb_driver: add a script to install Cypress usb driver
The new DVKs of the cc8mn, cc8mm and ccmp1 have a new ftdi
usb to serial chip that is recognized as a thermal device by default.

With the install_usb_driver.sh script this driver is replaced
to a USB to serial driver.

This script is included for the needed platforms in the zip
installer provided in the getting started.

https://onedigi.atlassian.net/browse/DEL-8126

Signed-off-by: Francisco Gil <francisco.gilmartinez@digi.com>