imx-boot, imx-mkimage: update to version lf-5.15.71-2.2.0
Both recipes (imx-boot and imx-mkimage) use the same source code
repository, so update both of them to the new NXP release, by copying
with minimal changes the recipes in meta-imx.

Also convert the original 'imx-mkimage' recipe, which was only for
native class, to a full target, native, nativesdk recipe. We need this
to be able to include the nativesdk one in our toolchain.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>