ccimx93: imx-boot: generate properly named boot artifacts
Although in theory, you can use any label to name the different u-boot
build configurations, we have them coupled to machine names in the
boot-artifacts bbclass, and also in the default boot artifact filename
in the firmware installation scripts.

So fix that up for the ccimx93 in the machine config, and create the
proper symlinks in the do_deploy imx-boot recipe.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>