bluez5: v5.65: update Digi customizations patch to remove battery plugin
Starting at Bluez 5.48, the battery characteristic was moved to the
DBUS org.bluez.Battery1 interface. This causes the device to try to
read information from iOS devices after establishing a connection,
triggering a reverse pairing request. This scenario causes random
disconnects in iOS devices unless a trust agent is registered in the
host to take care of the pairing. Removing the battery plugin at
startup fixes the issue.

Signed-off-by: David Escalona <david.escalona@digi.com>