dey-image-graphical: add 'weston' image feature
For 'wayland' enabled distro features, 'weston' is added as image
feature. This is later used in poky to change the SYSTEMD_DEFAULT_TARGET
to 'graphical.target' which is the correct systemd default target for
graphical images.

This allows to delete the workaround we have in 'weston-init'
recipe to start weston automatically on boot.

Also delete some packages from CORE_IMAGE_EXTRA_INSTALL:

- weston-init and weston-examples are included as part of the 'weston'
  image feature.
- gtk+3-demo: removed from the default images, as it is a graphic framework
  we are not promoting.
- xterm: removed as weston does already provide a terminal (weston-terminal).

Signed-off-by: Javier Viguera <javier.viguera@digi.com>