meta-digi: get rid of accel-video machine feature
accel-video was being used only for including or not gstreamer in the
different DEY images. But then we had exceptions to include gstreamer for
machines that do not define accel-video, so just follow what the ccmp15
does, and use a generic 'gstreamer' distro feature to add 'dey-gstreamer'
support to the images.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>