packagegroup-dey-gstreamer: rework MACHINE_GSTREAMER_1_0_EXTRA_INSTALL
We need some of the extra gstreamer packages added via
MACHINE_GSTREAMER_1_0_EXTRA_INSTALL for the ccimx93, but as the ccimx93
does not have an 'imxgpu' override, they don't get added. Instead, add
the packages for all the NXP platforms by removing the override and then
empty the variable for the 'ccimx6ul' for space-saving purposes.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>